+++ R/alpha_div_test.R
@@ -135,7 +135,7 @@ hill_tuckey_pq <- function(
 #'   - "robust"
 #'   - "bayes"
 #'
-#' @param ... Other arguments passed on to [ggstatsplot::ggbetweenstats()] function
+#' @param ... Additional arguments passed on to [ggstatsplot::ggbetweenstats()] function
 #' @seealso [ggstatsplot::ggbetweenstats()], [hill_pq()]
 #' @return A list of 6 components :
 #'
@@ -305,7 +305,7 @@ hill_test_rarperm_pq <- function(physeq,
 #' @param crit The Information Criterion to be used. Default is the small-sample corrected AIC (aicc).
 #'   This should be a function that accepts a fitted model as first argument.
 #'   Other provided functions are the classic AIC, the Bayes IC (bic), and QAIC/QAICc (qaic and qaicc).
-#' @param ... Other arguments passed on to [glmulti::glmulti()] function
+#' @param ... Additional arguments passed on to [glmulti::glmulti()] function
 #'
 #' @return A data.frame summarizing the glmulti results with columns
 #'