+++ R/vsearch.R
@@ -507,7 +507,7 @@ vsearch_clustering <- function(physeq = NULL,
 #' @param clean_pq (logical; default FALSE) If TRUE, return the phyloseq object
 #'   after cleaning using the default parameter of [clean_pq()] function.
 #'
-#' @param ... Other arguments passed on to [chimera_detection_vs()] function
+#' @param ... Additional arguments passed on to [chimera_detection_vs()] function
 #' @seealso [chimera_detection_vs()], [dada2::removeBimeraDenovo()]
 #' @return
 #'
@@ -855,11 +855,11 @@ write_temp_fasta <- function(physeq,
 #' @param nproc (default: 1)
 #'   Set to number of cpus/processors to use
 #' @param suffix (character) The suffix to name the new columns.
-#'   If set to "" (the default), the taxo_rank algorithm is used
+#'   If set to "" (the default), the taxa_ranks algorithm is used
 #'   without suffix.
-#' @param taxo_rank A list with the name of the taxonomic rank present in
+#' @param taxa_ranks A list with the name of the taxonomic rank present in
 #'   ref_fasta
-#' @param min_bootstrap (Int. \[0:1\], default 0.5)
+#' @param min_bootstrap (Float \[0:1\], default 0.5)
 #'   Minimum bootstrap value to inform taxonomy. For each bootstrap
 #'   below the min_bootstrap value, the taxonomy information is set to NA.
 #' @param keep_temporary_files (logical, default: FALSE) Do we keep temporary files?
@@ -871,7 +871,7 @@ write_temp_fasta <- function(physeq,
 #'
 #' @param verbose (logical). If TRUE, print additional information.
 #' @param temporary_fasta_file The name of a temporary_fasta_file (default "temp.fasta")
-#' @param cmd_args Other arguments passed on to vsearch sintax cmd.
+#' @param cmd_args Additional arguments passed on to vsearch sintax cmd.
 #'   By default cmd_args is equal to "--sintax_random" as recommended by
 #'   [Torognes](https://github.com/torognes/vsearch/issues/535).
 #' @param too_few (default value "align_start") see [tidyr::separate_wider_delim()]
@@ -927,7 +927,7 @@ assign_sintax <- function(physeq = NULL,
                           clean_pq = TRUE,
                           nproc = 1,
                           suffix = "",
-                          taxo_rank = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
+                          taxa_ranks = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
                           min_bootstrap = 0.5,
                           keep_temporary_files = FALSE,
                           verbose = TRUE,
@@ -991,7 +991,7 @@ assign_sintax <- function(physeq = NULL,
   taxa_names <- res_sintax$V1
   res_sintax <- tibble(res_sintax$V2, taxa_names)
   res_sintax <- res_sintax |>
-    tidyr::separate_wider_delim(-taxa_names, names = paste0(taxo_rank, suffix), delim = ",", too_few = too_few) |>
+    tidyr::separate_wider_delim(-taxa_names, names = paste0(taxa_ranks, suffix), delim = ",", too_few = too_few) |>
     tidyr::pivot_longer(-taxa_names) |>
     tidyr::separate_wider_delim(
       value,
@@ -1090,14 +1090,14 @@ assign_sintax <- function(physeq = NULL,
 #' @param clean_pq (logical, default TRUE)
 #'   If set to TRUE, empty samples and empty ASV are discarded
 #'   before clustering.
-#' @param taxo_rank A list with the name of the taxonomic rank present in
+#' @param taxa_ranks A list with the name of the taxonomic rank present in
 #'   ref_fasta
 #' @param nproc (int, default: 1)
 #'   Set to number of cpus/processors to use
 #' @param suffix (character) The suffix to name the new columns.
-#'   If set to "" (the default), the taxo_rank algorithm is used
+#'   If set to "" (the default), the taxa_ranks algorithm is used
 #'   without suffix.
-#' @param id (Int. \[0:1\] default 0.5). Default value is based on
+#' @param id (Float \[0:1\] default 0.5). Default value is based on
 #'   [stampa](https://github.com/frederic-mahe/stampa).
 #'   See Vsearch Manual for parameter `--id`
 #' @param lca_cutoff (int, default 1). Fraction of matching hits
@@ -1144,7 +1144,7 @@ assign_sintax <- function(physeq = NULL,
 #' @param verbose (logical). If TRUE, print additional information.
 #' @param temporary_fasta_file Name of the temporary fasta file. Only useful
 #'   with keep_temporary_files = TRUE.
-#' @param cmd_args Other arguments passed on to vsearch usearch_global cmd.
+#' @param cmd_args Additional arguments passed on to vsearch usearch_global cmd.
 #' @param too_few (default value "align_start") see [tidyr::separate_wider_delim()]
 #' @return See param behavior
 #' @seealso [assign_sintax()], [add_new_taxonomy_pq()]
@@ -1167,7 +1167,7 @@ assign_vsearch_lca <- function(physeq = NULL,
                                behavior = c("return_matrix", "add_to_phyloseq", "return_cmd"),
                                vsearchpath = "vsearch",
                                clean_pq = TRUE,
-                               taxo_rank = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
+                               taxa_ranks = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
                                nproc = 1,
                                suffix = "",
                                id = 0.5,
@@ -1241,7 +1241,7 @@ assign_vsearch_lca <- function(physeq = NULL,
 
   res_usearch <- res_usearch |>
     tidyr::separate_wider_delim(-taxa_names,
-      names = paste0(taxo_rank, suffix),
+      names = paste0(taxa_ranks, suffix),
       delim = ",",
       too_few = too_few
     ) |>